Fix a harmless MSVC compiler warning.
FossilOrigin-Name: 6cc023bb29be51847fbbfab95c24fc89993ccdba

+++ ext/fts3/fts3.c
@@ -360,7 +360,7 @@ int sqlite3Fts3GetVarint(const char *p, sqlite_int64 *v){
     if( (c & 0x80)==0 ) break;
   }
   *v = b;
-  return p - pStart;
+  return (int)(p - pStart);
 }
 
 /*
